feat: extract logprob and tensors structures
